Add files via upload
This is an autosplitter for Wrath: Aeon of Ruin. There were some versions before this, but this is the first one I'm uploading to GitHub. The game is in early access, so this autosplitter is likely to become outdated at some point whenever the game gets updated, but it doesn't seem to be happening often, so perhaps it might be a while before I have to make changes. At the time of pushing this first version, the game has 5 levels plus the main hub.